+++ app/src/organisms/Desktop/ProtocolDetails/AnnotatedSteps/ProtocolAnalysisErrorsContent.tsx
@@ -37,9 +37,7 @@ export function ProtocolAnalysisErrorsContent(
         <div
           className={styles.annotated_steps_error_container}
           key={error.id}
-          onClick={() => {
-            onShowErrorDetails()
-          }}
+          onClick={onShowErrorDetails}
         >
           <div className={styles.annotated_steps_header}>
             <Icon name="ot-alert" size="1rem" color={COLORS.red60} />
@@ -63,18 +61,3 @@ export function ProtocolAnalysisErrorsContent(
   )
 }
 
-export function ProtocolAnalysisPastStepsMessage(props: {
-  t: (key: string) => string
-}): JSX.Element {
-  const { t } = props
-
-  return (
-    <div className={styles.annotated_steps_past_steps_row}>
-      <div className={styles.annotated_steps_error_footer_message}>
-        <StyledText desktopStyle="bodyDefaultRegular" color={COLORS.grey60}>
-          {t('unable_to_show_steps_past_errors')}
-        </StyledText>
-      </div>
-    </div>
-  )
-}
