+++ protocol-visualization/src/organisms/AnnotatedSteps/AnnotatedStepsRowItem.tsx
@@ -6,9 +6,9 @@ import { AnnotatedGroup } from './AnnotatedGroup'
 import styles from './annotatedsteps.module.css'
 import { IndividualCommand } from './IndividualCommand'
 import {
-  ProtocolAnalysisErrorsContent,
-  ProtocolAnalysisPastStepsMessage,
+  ProtocolAnalysisErrorsContent
 } from './ProtocolAnalysisErrorsContent'
+import { ProtocolAnalysisPastStepsMessage } from './ProtocolAnalysisPastStepsMessage'
 
 import type { Dispatch, SetStateAction } from 'react'
 import type { RowComponentProps } from 'react-window'
